+++ source/Public/Initialize-TestEnvironment.ps1
@@ -38,6 +38,16 @@
         Mof: The test initialization assumes a Mof-based DscResource folder structure.
         Class: The test initialization assumes a Class-based DscResource folder structure.
 
+    .PARAMETER ProcessExecutionPolicy
+        Specifies the process execution policy to set before running tests.
+        The default will be the one that is set on the machine where the tests
+        are running.
+
+    .PARAMETER MachineExecutionPolicy
+        Specifies the machine execution policy to set before running tests.
+        The default will be the one that is set on the machine where the tests
+        are running.
+
     .EXAMPLE
         $TestEnvironment = Initialize-TestEnvironment `
             -DSCModuleName 'xNetworking' `
@@ -91,7 +101,17 @@ function Initialize-TestEnvironment
         [Parameter()]
         [ValidateSet('Mof', 'Class')]
         [String]
-        $ResourceType = 'Mof'
+        $ResourceType = 'Mof',
+
+        [Parameter()]
+        [ValidateSet('AllSigned', 'Bypass','RemoteSigned','Unrestricted')]
+        [String]
+        $ProcessExecutionPolicy,
+
+        [Parameter()]
+        [ValidateSet('AllSigned', 'Bypass','RemoteSigned','Unrestricted')]
+        [String]
+        $MachineExecutionPolicy
     )
 
     Write-Verbose -Message "Initializing test environment for $TestType testing of $DscResourceName in module $Module"
@@ -200,14 +220,26 @@ function Initialize-TestEnvironment
             }
 
             # Preserve and set the execution policy so that the DSC MOF can be created
-            if (!$script:MachineOldExecutionPolicy)
+            $currentMachineExecutionPolicy = Get-ExecutionPolicy -Scope LocalMachine
+            if ($PSBoundParameters.ContainsKey('MachineExecutionPolicy'))
             {
-                $script:MachineOldExecutionPolicy = Get-ExecutionPolicy -Scope LocalMachine
-                if ($script:MachineOldExecutionPolicy -ine 'Unrestricted')
+                if ($currentMachineExecutionPolicy -ne $MachineExecutionPolicy)
                 {
-                    Set-ExecutionPolicy -ExecutionPolicy 'Unrestricted' -Scope LocalMachine -Force -ErrorAction Stop
+                    Set-ExecutionPolicy -ExecutionPolicy $MachineExecutionPolicy -Scope 'LocalMachine' -Force -ErrorAction Stop
+
+                    <#
+                        Should only be set after we actually changed the execution
+                        policy because if $script:MachineOldExecutionPolicy is set
+                        to a value `Restore-TestEnvironment` will try to revert
+                        the value.
+                    #>
+                    $script:MachineOldExecutionPolicy = $currentMachineExecutionPolicy
+
+                    $currentMachineExecutionPolicy = $MachineExecutionPolicy
                 }
             }
+
+            Write-Verbose -Message ('The machine execution policy is set to ''{0}''' -f $currentMachineExecutionPolicy)
 
             <#
                 For integration tests we have to set the machine's PSModulePath because otherwise the
@@ -227,12 +259,25 @@ function Initialize-TestEnvironment
         }
     }
 
-    # Preserve and set the execution policy so that the DSC MOF can be created
-    $oldExecutionPolicy = Get-ExecutionPolicy
-    if ($oldExecutionPolicy -ine 'Unrestricted')
+    <#
+        Preserve and set the execution policy so that the DSC MOF can be created.
+
+        `Restore-TestEnvironment` will only revert the value if $oldExecutionPolicy
+        differ from current execution policy. So we make to always set it to the
+        current execution policy so that if we don't need to change it then
+        `Restore-TestEnvironment` will not try to revert the value.
+    #>
+    $oldExecutionPolicy = Get-ExecutionPolicy -Scope 'Process'
+    if ($PSBoundParameters.ContainsKey('ProcessExecutionPolicy'))
     {
-        Set-ExecutionPolicy -ExecutionPolicy 'Unrestricted' -Scope 'Process' -Force
+        if ($oldExecutionPolicy -ne $ProcessExecutionPolicy)
+        {
+            Set-ExecutionPolicy -ExecutionPolicy $ProcessExecutionPolicy -Scope 'Process' -Force -ErrorAction Stop
+        }
     }
+
+    Write-Verbose -Message ('The process execution policy is set to ''{0}''' -f $currentProcesoldExecutionPolicysExecutionPolicy)
+
 
     # Return the test environment
     return @{